Refactor component's folder structure under src/screens/Users
