+++ src/screens/Users/screens/SignUp/components/Form/index.jsx
@@ -20,11 +20,11 @@ import PATHS from 'constants/router/paths';
 
 // Action creators and selectors
 import { signUp } from 'data/session/actions';
-import { resetUI } from '../actions';
-import { getUI } from '../reducers';
+import { resetUI } from '../../actions';
+import { getUI } from '../../reducers';
 
 // Companion files
-import FIELDS from '../constants/fields';
+import FIELDS from '../../constants/fields';
 
 // Component
 class SignUpForm extends Component {
